move help commands to archiver.help
